[9] New Core Revision 9 Database coming soon

+++ src/server/game/Server/WorldSocket.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2008-2010 TrinityCore <http://www.trinitycore.org/>
+ * Copyright (C) 2008-2011 TrinityCore <http://www.trinitycore.org/>
  * Copyright (C) 2005-2009 MaNGOS <http://getmangos.com/>
  *
  * This program is free software; you can redistribute it and/or modify it
@@ -130,7 +130,7 @@ class WorldSocket : protected WorldHandler
         virtual int open (void *);
 
         /// Called on failures inside of the acceptor, don't call from your code.
-        virtual int close (int);
+        virtual int close (u_long);
 
         /// Called when we can read from the socket.
         virtual int handle_input (ACE_HANDLE = ACE_INVALID_HANDLE);
